+++ Solution/src/main/java/com/example/demo/repositories/TransactionRepository.java
@@ -4,7 +4,6 @@ import com.example.demo.models.transaction.Deposit;
 import com.example.demo.models.transaction.Internal;
 import com.example.demo.models.transaction.Transaction;
 import com.example.demo.models.transaction.Withdrawal;
-import org.hibernate.Session;
 
 import java.util.List;
 
@@ -13,11 +12,11 @@ public interface TransactionRepository {
 
     List<Transaction> getTransactionsbyWalletId(int id);
 
-    Internal createInternal(Internal internal, int balanceSender, int balanceReceiver, int senderId, int receiverId);
+    Internal createInternal(Internal internal, double balanceSender, double balanceReceiver, int senderId, int receiverId);
 
-    Withdrawal createWithdrawal(Withdrawal withdrawal, int balanceSender, int senderId);
+    Withdrawal createWithdrawal(Withdrawal withdrawal, double balanceSender, int senderId);
 
-    Deposit createDeposit(Deposit deposit, int balanceReceiver, int receiverId);
+    Deposit createDeposit(Deposit deposit, double balanceReceiver, int receiverId);
 
     boolean checkIfIdempotencyKeyExists(String idempotencyKey);
